requirements.txt
openai==1.109.1
numpy
pandas
azure-identity
azure-search-documents
python-dotenv
pillow
promptflow-evals
ipykernel